replymate_core/replymate-core/requirements.txt

fastapi
uvicorn
google-auth-oauthlib
google-api-python-client
openai
pydantic
